projeto/qual-a-boa-backend/db-api/requirements.txt
# Framework web e suas extensões
Flask==2.3.3  # Framework web principal
Flask-SQLAlchemy==3.1.1  # ORM para banco de dados
Flask-Migrate==4.0.5  # Gerenciamento de migrações do banco de dados
Flask-JWT-Extended==4.4.4  # Autenticação via JWT
Flask-CORS==4.0.0  # Suporte a CORS
Flask-WTF==1.2.1  # Proteção CSRF

# Dependências principais
Werkzeug==2.3.7  # WSGI utilities
bcrypt==4.0.1  # Hashing de senhas
python-dotenv==1.0.0  # Gerenciamento de variáveis de ambiente

# Servidor WSGI para produção
gunicorn==21.2.0

# Testes
pytest==7.4.2  # Framework de testes
pytest-flask==1.2.0  # Extensão para testes Flask

# Validação de dados
marshmallow==3.20.1  # Serialização e validação de dados 